pyproject.toml
[tool.poetry]
name = "local-delivery"
version = "0.1.0"
description = ""
authors = ["Vitaliy Mishkov <vitaliy.mishkov@pm.by>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.17.1"
python-dotenv = "^0.19.2"
aioredis = "^2.0.1"
asyncpg = "^0.25.0"
emoji = "^1.6.1"
XlsxWriter = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
